Refactor to allow integration choices
So far there are 2 different databases supported,
postgresql as well as the old mariadb/mysql.

The plan is to allow different combinations of app server
and web server to serve redmine, but so far only one option
has been fully implemented and tested (uwsgi as app server
with uwsgi also as web server).