connect to postgresql with unix socket for now
For some reason postgresql IP socket on localhost is not working on the
centos7-dev VM I'm trying to set up

  rake aborted!
  PG::ConnectionBad: FATAL:  Ident authentication failed for user "redmine"

even though /var/lib/pgsql/data/pg_hba.conf was the same (left at default)
and this role is supposed to be setting the redmine database password.

Same user works fine with "peer" authentication over a unix socket.
Go with that for now, since database is running on the same host.